requirements.txt
google-adk
streamlit
requests
python-dotenv
